backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
adbc-driver-manager==0.5.1
    # via adbc-driver-sqlite
adbc-driver-sqlite==0.5.1
    # via polars
annotated-types==0.5.0
    # via pydantic
apispec[yaml]==6.3.0
    # via flask-appbuilder
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
babel==2.12.1
    # via flask-babel
blinker==1.6.2
    # via -r requirements.in
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via
    #   flask
    #   flask-appbuilder
cloudpickle==2.2.1
    # via polars
colorama==0.4.6
    # via flask-appbuilder
connectorx==0.2.3
    # via polars
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
db-dtypes==1.1.1
    # via -r requirements.in
deltalake==0.10.1
    # via polars
deprecated==1.2.14
    # via limits
dnspython==2.4.2
    # via email-validator
email-validator==1.3.1
    # via flask-appbuilder
flask==2.2.5
    # via
    #   -r requirements.in
    #   flask-appbuilder
    #   flask-babel
    #   flask-cors
    #   flask-jwt-extended
    #   flask-limiter
    #   flask-login
    #   flask-sqlalchemy
    #   flask-wtf
flask-appbuilder==4.3.6
    # via -r requirements.in
flask-babel==2.0.0
    # via flask-appbuilder
flask-cors==4.0.0
    # via -r requirements.in
flask-jwt-extended==4.5.2
    # via flask-appbuilder
flask-limiter==3.5.0
    # via flask-appbuilder
flask-login==0.6.2
    # via flask-appbuilder
flask-sqlalchemy==2.5.1
    # via flask-appbuilder
flask-wtf==1.1.1
    # via flask-appbuilder
fonttools==4.42.1
    # via matplotlib
fsspec==2023.6.0
    # via polars
google-api-core[grpc]==2.11.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.22.0
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-bigquery==3.11.4
    # via -r requirements.in
google-cloud-core==2.3.3
    # via google-cloud-bigquery
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.5.0
    # via google-cloud-bigquery
googleapis-common-protos==1.60.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==2.0.2
    # via sqlalchemy
grpcio==1.56.2
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.56.2
    # via google-api-core
idna==3.4
    # via
    #   email-validator
    #   requests
importlib-resources==6.0.1
    # via limits
itsdangerous==2.1.1
    # via
    #   flask
    #   flask-wtf
jinja2==3.0.3
    # via
    #   flask
    #   flask-babel
jsonschema==4.19.0
    # via flask-appbuilder
jsonschema-specifications==2023.7.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
limits==3.6.0
    # via flask-limiter
loguru==0.7.0
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
marshmallow==3.20.1
    # via
    #   flask-appbuilder
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.26.1
    # via flask-appbuilder
matplotlib==3.7.2
    # via polars
mdurl==0.1.2
    # via markdown-it-py
numpy==1.25.1
    # via
    #   connectorx
    #   contourpy
    #   db-dtypes
    #   matplotlib
    #   pandas
    #   polars
    #   pyarrow
    #   scipy
ordered-set==4.1.0
    # via flask-limiter
orjson==3.9.5
    # via -r requirements.in
packaging==23.1
    # via
    #   apispec
    #   db-dtypes
    #   google-cloud-bigquery
    #   limits
    #   marshmallow
    #   matplotlib
pandas==2.0.3
    # via
    #   -r requirements.in
    #   db-dtypes
    #   polars
pillow==10.0.0
    # via matplotlib
polars[adbc,all,cloudpickle,connectorx,deltalake,fsspec,matplotlib,numpy,pandas,pyarrow,pydantic,sqlalchemy,timezone,xlsx2csv,xlsxwriter]==0.18.15
    # via -r requirements.in
prison==0.2.1
    # via flask-appbuilder
proto-plus==1.22.3
    # via google-cloud-bigquery
protobuf==4.23.4
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyarrow==12.0.1
    # via
    #   db-dtypes
    #   deltalake
    #   polars
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==2.3.0
    # via polars
pydantic-core==2.6.3
    # via pydantic
pygments==2.16.1
    # via rich
pyjwt==2.8.0
    # via
    #   flask-appbuilder
    #   flask-jwt-extended
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   flask-appbuilder
    #   google-cloud-bigquery
    #   matplotlib
    #   pandas
pytz==2021.3
    # via
    #   flask-babel
    #   pandas
pyyaml==6.0.1
    # via apispec
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
rich==13.5.2
    # via flask-limiter
rpds-py==0.10.2
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
scipy==1.11.2
    # via -r requirements.in
sentry-sdk==1.29.2
    # via -r requirements.in
six==1.16.0
    # via
    #   google-auth
    #   prison
    #   python-dateutil
sqlalchemy==1.4.49
    # via
    #   flask-appbuilder
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
    #   polars
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.1
    # via flask-appbuilder
typing-extensions==4.7.1
    # via
    #   flask-limiter
    #   limits
    #   pydantic
    #   pydantic-core
tzdata==2023.3
    # via pandas
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
    #   sentry-sdk
werkzeug==2.3.6
    # via
    #   flask
    #   flask-jwt-extended
    #   flask-login
wrapt==1.15.0
    # via deprecated
wtforms==3.0.1
    # via
    #   flask-appbuilder
    #   flask-wtf
xlsx2csv==0.8.1
    # via polars
xlsxwriter==3.1.2
    # via polars
